Replace termbox with tcell

+++ fluid-solver/particle.go
@@ -11,32 +11,32 @@ func NewParticle(x, y float64) *Particle {
 	return &Particle{x: x, y: y}
 }
 
-func (p *Particle) GetX() int {
-	return int(p.x)
+func (p *Particle) GetX() float64 {
+	return p.x
 }
 
 func (p *Particle) SetX(val float64) {
 	p.x = val
 }
 
-func (p *Particle) GetY() int {
-	return int(p.y)
+func (p *Particle) GetY() float64 {
+	return p.y
 }
 
 func (p *Particle) SetY(val float64) {
 	p.y = val
 }
 
-func (p *Particle) GetVx() int {
-	return int(p.vx)
+func (p *Particle) GetVx() float64 {
+	return p.vx
 }
 
 func (p *Particle) SetVx(val float64) {
 	p.vx = val
 }
 
-func (p *Particle) GetVy() int {
-	return int(p.vy)
+func (p *Particle) GetVy() float64 {
+	return p.vy
 }
 
 func (p *Particle) SetVy(val float64) {
